Refatorando Controller, Services e Repository

+++ src/main/java/br/com/hotelCalifornia/domain/service/HotelCaliforniaService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import java.util.logging.Logger;
 
 import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,11 +28,18 @@ public class HotelCaliforniaService {
     }
 	
 	
-	public List<HotelCaliforniaModel> findAll() {
+	public ResponseEntity<Object> findAll() {
 	
 		logger.info("Metodo findAll");
-         
-		return repository.findAll();
+		
+		List<HotelCaliforniaModel> listaTodos = repository.findAll();
+		
+		if(listaTodos.isEmpty()) {
+			return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("Hotel não encontrado");}
+		
+		return ResponseEntity.status(HttpStatus.OK).body(listaTodos); 
+		
+		
 		
 	}
 	@Transactional
@@ -42,41 +50,76 @@ public class HotelCaliforniaService {
 		return repository.save(hotelCaliforniaModel);
 		
 	}
-	public Optional<HotelCaliforniaModel> acharId(Long id) {
+	public ResponseEntity<Object> acharId(Long id) {
 		
     		logger.info("Metodo acharId");
-	
-	    	return repository.findById(id);
+    		
+    		Optional<HotelCaliforniaModel> californiaModel = repository.findById(id);
+    		if(!californiaModel.isPresent()) 
+        		return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("Hotel não encontrado");
+			return ResponseEntity.status(HttpStatus.OK).body(californiaModel);
 	}
-	public ResponseEntity<HotelCaliforniaModel> buscarPorCnpj(String cnpj) {
+	public ResponseEntity<Object> buscarPorCnpj(String cnpj) {
 		
 	     	logger.info("Metodo acharPorCNPJ");
 	     	
-	        return repository.acharPorCnpj(cnpj).map(mapping->ResponseEntity.ok().body(mapping))
-	    			.orElse(ResponseEntity.notFound().build());
+	    	Optional<HotelCaliforniaModel> californiaModel =  repository.acharPorCnpj(cnpj);
+    		if(!californiaModel.isPresent()) 
+        		return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("Hotel não encontrado");
+    		
+    		return ResponseEntity.status(HttpStatus.OK).body(californiaModel);  
+	        
 	        }
+	public ResponseEntity<HotelCaliforniaModel> buscarPorlocal(String local) {
+		
+     	logger.info("Metodo acharPorLocal");
+     	
+        return repository.acharPorLocal(local).map(mapping->ResponseEntity.ok().body(mapping))
+    			.orElse(ResponseEntity.notFound().build());
+        }
 	@Transactional
-	public HotelCaliforniaModel atualizar(Long id, HotelCaliforniaModel hotelCaliforniaModel) {
-		logger.info("Metodo update");
-
+	public ResponseEntity<Object> atualizar(Long id, HotelCaliforniaModel hotelCaliforniaModel) {
+		logger.info("Metodo Atualizar");
+  
+        Optional<HotelCaliforniaModel> hotelOptional = repository.findById(id);
+    	
+    	if(!hotelOptional.isPresent()) {
+    		return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("Hotel não encontrado");}
 		
 		HotelCaliforniaModel novoHotel = repository.findById(id).get();
     	BeanUtils.copyProperties(hotelCaliforniaModel, novoHotel,"id"); // o terceiro parametro "id" assegura a alteração do registro
     	                                                                // se não colocar, a biblioteca vai criar um novo registro com um
     	                                                               // com um novo id, com os dados alterados  
     	
-    	return repository.save(novoHotel);          			
+    	try {
+    		repository.save(novoHotel);
+	        return ResponseEntity.status(HttpStatus.OK).body(novoHotel);  
+			
+		} catch (Exception e) {
+			return  ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Erro ! Hotel não atualizado");
+		}
+    	       			
 	  }
 	 @Transactional
 	 public ResponseEntity<?> deletar(@PathVariable Long id) {
 		 
 		logger.info("Metodo delete");
+		Optional<HotelCaliforniaModel> hotelOptional = repository.findById(id);
+	    	
+	    if(!hotelOptional.isPresent()) {
+	    		return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("Hotel não encontrado");}
+			
+	    try {
+	    	return repository.findById(id).map(mapping->{
+	     		   repository.deleteById(id);
+	      	
+	     	       return ResponseEntity.ok().body("DELETADO COM SUCESSO");}
+	         ).orElse(ResponseEntity.noContent().build());   
+		} catch (Exception e) {
+			return  ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Erro ! Hotel não deletado");
+		}
  
-     	return repository.findById(id).map(mapping->{
-     		   repository.deleteById(id);
-      	
-     	       return ResponseEntity.ok().body("DELETADO COM SUCESSO");}
-         ).orElse(ResponseEntity.notFound().build());   
+     	
 
  }
